Convert trials to Int rather than Int64 for Mimi
Mimi's run method for a SimulationDef takes an Int, which is the
platform's native width, so the previous Int64 conversion moved the
32-bit Windows failure from run_mcs's own signature into that call
instead of fixing it. Int is Int32 on a 32-bit platform and Int64 on a
64-bit one, so converting to Int works on both.

+++ src/mcs/AR6_Monte_Carlo.jl
@@ -28,10 +28,11 @@ function run_mcs(;trials::Integer = 2237,
     # Get an instance of the mcs
     mcs = get_mcs()
 
-    # run monte carlo trials, converting to Int64 because that is what Mimi's
-    # run method for a SimulationDef takes. On a 32-bit platform an integer
-    # literal is an Int32, so without this a caller could not pass one at all.
-    results = run(mcs, m, Int64(trials); trials_output_filename = trials_output_filename, results_output_dir = "$output_dir/results")
+    # run monte carlo trials, converting to Int because that is what Mimi's run
+    # method for a SimulationDef takes. Int is the platform's native width, so
+    # this is Int32 on a 32-bit platform and Int64 on a 64-bit one; hard-coding
+    # either would break the other.
+    results = run(mcs, m, Int(trials); trials_output_filename = trials_output_filename, results_output_dir = "$output_dir/results")
 
     return results
 end
